+++ Sources/SelectableText/Extensions/SelectableTextRepresentable+UIKit.swift
@@ -12,10 +12,20 @@ import SwiftUI
 struct SelectableTextRepresentable: UIViewRepresentable {
     var text: String? = nil
     var attributedText: NSAttributedString? = nil
-    var menuBuilder: ((UIMenu) -> UIMenu)? = nil
+    var menuBuilder: (([UIMenuElement], UITextView) -> UIMenu)? = nil
     
     var maxLayoutWidth: CGFloat = .zero
     @Binding var layoutHeight: CGFloat
+  
+    func makeCoordinator() -> MenuCoordinator {
+        MenuCoordinator { suggested, textView in
+          if let menuBuilder = menuBuilder {
+             return menuBuilder(suggested, textView)
+          } else {
+            return UIMenu(children: suggested)
+          }
+        }
+    }
     
     func makeUIView(context: Context) -> UITextView {
         let textView = BaseTextView()
@@ -38,8 +48,10 @@ struct SelectableTextRepresentable: UIViewRepresentable {
             textView.attributedText = attributedText
         }
 
-        if #available(iOS 16, *), let builder = menuBuilder {
-            textView.editMenuProvider = { suggested in builder(suggested) }
+        if #available(iOS 16, *) {
+            if menuBuilder != nil {
+                textView.delegate = context.coordinator
+            }
         }
         
         return textView